Deprecates Endpoint.create(String,int,int) as it leads to null pointers
Multiple people have been bitten by `Endpoint.create(String,int,int)` as
`Endpoint.port` is nullable. This deprecates that method and advises to
use a better documented builder method or simply `Endpoint.toBuilder()`
when copying an endpoint.

+++ zipkin/src/main/java/zipkin/internal/JsonCodec.java
@@ -81,7 +81,7 @@ public final class JsonCodec implements Codec {
           // Shouldn't hit DNS, because it's an IP string literal.
           result.ipv6(Inet6Address.getByName(input).getAddress());
         } else if (nextName.equals("port")) {
-          result.port((short) reader.nextInt());
+          result.port(reader.nextInt());
         } else {
           reader.skipValue();
         }
